Rename: purchase, deposit으로 변경, roomAction, chatAction으로 변경
