+++ client/src/_actions/type.js
@@ -13,11 +13,11 @@ export const LOGOUT_USER = "logout";
 export const AUTH_USER = "auth";
 //product
 export const ADD_PRODUCT = "addProduct";
-export const PURCHASED_PRODUCTS = "purchasedProducts";
+export const DEPOSITED_PRODUCTS = "depositedProducts";
 export const GET_PRODUCTS = "getProducts";
 export const RECENT_PRODUCTS = "recentProducts";
 export const GET_PRODUCT = "getProduct";
-export const PURCHASE = "purchase";
+export const DEPOSIT = "deposit";
 export const RELEASE = "release";
 //ui
 export const SET_LOADINGS = "setLoadings";
